restructure init func
